add total contract tax to contract stats

+++ persist/postgres/init.sql
@@ -565,13 +565,13 @@ CREATE TABLE sectors (
 
 CREATE TABLE stats (
     stat_name TEXT PRIMARY KEY NOT NULL,
-    stat_value BIGINT NOT NULL DEFAULT 0 CHECK (stat_value >= 0)
+    stat_value NUMERIC(50,0) NOT NULL DEFAULT 0 CHECK (stat_value >= 0)
 );
 
 CREATE TABLE stats_deltas (
     id BIGSERIAL PRIMARY KEY,
     stat_name TEXT NOT NULL REFERENCES stats(stat_name),
-    stat_delta BIGINT NOT NULL
+    stat_delta NUMERIC(50,0) NOT NULL
 );
 
 -- speed up summing pending stats deltas for a particular stat
